requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Pydantic & settings
pydantic==2.7.1
pydantic-settings==2.2.1

# Database
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1

# Redis
redis==5.0.4

# HTTP client
httpx==0.27.0

# LangGraph / LangChain
langgraph==0.0.55
langchain==0.2.1
langchain-openai==0.1.7
langchain-anthropic==0.1.13

# GitHub
PyGithub==2.3.0

# Config
python-dotenv==1.0.1

# UI
streamlit==1.35.0

# Testing
pytest==8.2.1
pytest-asyncio==0.23.7
httpx==0.27.0

# Utilities
python-multipart==0.0.9

# Code analysis
radon==6.0.1
semgrep==1.72.0
